Migrate all synchronous network calls from 'requests' to 'httpx'
Change variable name in '_download_async' method from
'max_num_coroutines' to 'max_connections' to comply to the docstring's
content. Change 'max_connections' from positional argument to keyword
argument for the same reason.

Change 'monitor_url' variable assignment in 'copy_item' method to the
value of the 'Location' HTTP header field of the response. This is
needed to be able to track the copy operation progress if needed.

Fix a small typo in '_create_headers' method's docstring.