Remove Support for Python 3.7 (#50)
Python 3.7 has reached "end of life". As such dependencies have dropped support for it which means that Graph OneDrive tests are now failing.
This pull request removes the support requirement for Python 3.7 as well as the tests for it.
This version should still work on 3.7 but in the future this may be broken without notice.
Release 0.4.0 is the last release to formally support Python 3.7.